requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-experimental>=0.3.0
langgraph>=0.2.0
openai>=1.50.0
faiss-cpu>=1.8.0
streamlit>=1.38.0
python-dotenv>=1.0.0
pypdf>=4.0.0
docx2txt>=0.8
unstructured>=0.15.0
openpyxl>=3.1.0
tiktoken>=0.7.0
